refactor(carve S7): remove graph CREATION from the reconciler, keep graph EXECUTION
With decomposition staying on the development branch, the reconciler no longer
needs to build a graph — a graph on main is human-authored and seeded by the
webhook. This removes the plan-consuming half and keeps the executing half.

Gone: the planning-task event shape and its parser, the plan reconciler, the
artifact fetch, the :auto seed path, and the seed-time attachment hydration. 656
contiguous lines, plus the six imports of the deleted planning modules and the
stream-handler branch that dispatched to them.

Kept, and worth naming because the file is now much smaller: terminal-child
reconciliation, dependent release, the restack cascade, panel refresh and settle,
failed-node recovery, and the iteration reply. Those are what the slice is for.

The removal also made three things dead that were only ever used by seeding — the
S3 client, the Bedrock screening config, and the artifacts-bucket env — so they
went too. Net effect on IAM is a least-privilege improvement: nothing in this
slice reads S3 any more.

Two consumers needed repointing rather than deleting, since they were never about
decomposition:

- `orchestration-rollup` imported DEFAULT_LABEL_FILTER from the deleted label
  module; it now takes it from `trigger-label`.
- A test asserted that a planning task falls through `parseTerminalTaskRecord`.
  The property it was really pinning is that the gate is the presence of
  orchestration_id, not the workflow id — a task outside a graph must not be
  mis-gated as a child and released against a graph it has nothing to do with.
  Rewritten around a plain terminal record so it still guards that.

Also drops a probe mock that existed because the seed path failed closed on an
attachment probe. With the seed gone the reconciler never probes, so the mock was
asserting nothing.
