+++ cdk/src/handlers/orchestration-reconciler.ts
@@ -36,9 +36,7 @@
  * terminal event neither double-releases nor regresses state.
  */
 
-import { BedrockRuntimeClient } from '@aws-sdk/client-bedrock-runtime';
 import { DynamoDBClient } from '@aws-sdk/client-dynamodb';
-import { GetObjectCommand, S3Client } from '@aws-sdk/client-s3';
 import {
   BatchGetCommand,
   DynamoDBDocumentClient,
@@ -47,26 +45,13 @@ import {
   UpdateCommand,
 } from '@aws-sdk/lib-dynamodb';
 import type { DynamoDBBatchResponse, DynamoDBRecord, DynamoDBStreamEvent } from 'aws-lambda';
-import type { ScreeningConfig } from './shared/attachment-screening';
 import { createTaskCore } from './shared/create-task-core';
 import { renderFailureReply, renderPanelFailureReason } from './shared/failure-reply';
 import { isNoChangeIteration, renderMaturingReply } from './shared/iteration-reply';
 import { claimTerminalReply, releaseReplyClaim } from './shared/iteration-reply-claim';
-import { downloadScreenAndStoreLinearAttachments, isLinearUploadsUrl, LinearAttachmentError } from './shared/linear-attachments';
-import { probeLinearIssueContext } from './shared/linear-issue-context-probe';
-import { resolveLinearOauthToken } from './shared/linear-oauth-resolver';
-import type { SubIssueNode } from './shared/linear-subissue-fetch';
 import { logger } from './shared/logger';
 import type { Channel, IssueRef } from './shared/orchestration-channel';
 import { channelForSource, type ChannelRegistryTables } from './shared/orchestration-channel-factory';
-import { applyDecompositionResult } from './shared/orchestration-decomposition-flow';
-import { parseDecomposerResponse } from './shared/orchestration-decomposition-planner';
-import { renderApprovedPlanReference, renderDecomposeUnavailableNote, renderRevisionToSingleNote } from './shared/orchestration-decomposition-render';
-import { getPendingPlan, putPendingPlan, replacePendingPlan } from './shared/orchestration-decomposition-store';
-import type { PlannedSubIssue, ProjectDecompositionCaps } from './shared/orchestration-decomposition-types';
-import { linearGraphqlFn } from './shared/orchestration-decomposition-writeback';
-import { discoverOrchestration } from './shared/orchestration-discovery';
-import { declarativeGraphSource } from './shared/orchestration-graph-source';
 import { isIntegrationNode } from './shared/orchestration-integration-node';
 import { ORCH_LOG } from './shared/orchestration-log-events';
 import {
@@ -79,15 +64,12 @@ import { applyTerminalCreateFailures, readConcurrencyBudget, releaseReadyChildre
 import { planDirectRestack, type RestackStep } from './shared/orchestration-restack';
 import { cascadeNodeLabel, upsertEpicPanel } from './shared/orchestration-rollup';
 import {
-  claimCommentAck,
   claimRollup,
   clearRollupClaim,
-  deriveOrchestrationId,
   loadOrchestration,
   setRetryCommentId,
   setStatusCommentId,
   type OrchestrationChildRow,
-  type OrchestrationReleaseContext,
 } from './shared/orchestration-store';
 import { encodeMarkdownUrl } from './shared/screenshot-url';
 import { OrchestrationTable } from '../constructs/orchestration-table';
@@ -123,10 +105,6 @@ const CHANNEL_REGISTRY_TABLES: ChannelRegistryTables = {
 const channelForMeta = (meta: { release_context?: { channel_source?: string } }): Channel | undefined =>
   channelForSource(meta.release_context?.channel_source, CHANNEL_REGISTRY_TABLES);
 
-/** Adapter for a surface named directly (a decompose event carries its own
- *  channel rather than a loaded orchestration row). */
-const channelFor = (source: string | undefined) => channelForSource(source, CHANNEL_REGISTRY_TABLES);
-
 /** Build the neutral issue reference the channel operates on. */
 const issueRef = (issueId: string, workspaceId: string): IssueRef => ({
   issueId,
@@ -140,34 +118,6 @@ const MAX_IDEMPOTENCY_KEY_LENGTH = 128;
 // table → no throttle (release-all, back-compat; admission still gates).
 const USER_CONCURRENCY_TABLE = process.env.USER_CONCURRENCY_TABLE_NAME;
 const MAX_CONCURRENT = Number(process.env.MAX_CONCURRENT_TASKS_PER_USER ?? '10');
-// Agent-native planning: the artifacts bucket a coding/decompose-v1 task
-// wrote its plan JSON to (artifacts/{task_id}/result.md). Unset → the decompose
-// terminal branch can't read plans and logs+skips (defensive; the construct
-// wires this alongside the read grant).
-const ARTIFACTS_BUCKET = process.env.ARTIFACTS_BUCKET_NAME;
-// Attachment hydration at decompose-seed time so every child inherits
-// the parent's screened attachments. Same env the webhook processor uses; unset →
-// the seed skips attachment hydration (children run without them, as before).
-const ATTACHMENTS_BUCKET = process.env.ATTACHMENTS_BUCKET_NAME;
-const GUARDRAIL_ID = process.env.GUARDRAIL_ID;
-const GUARDRAIL_VERSION = process.env.GUARDRAIL_VERSION;
-// TTL for a persisted pending plan awaiting @bgagent approve — mirrors the
-// webhook's PENDING_PLAN_TTL_SECONDS (a week).
-const PENDING_PLAN_TTL_SECONDS = 604_800;
-let sharedS3: S3Client | undefined;
-function s3(): S3Client {
-  if (!sharedS3) sharedS3 = new S3Client({});
-  return sharedS3;
-}
-let sharedAttachmentsBedrock: BedrockRuntimeClient | undefined;
-/** Lazily-built screening config for seed-time attachment hydration; null when
- *  the bucket/guardrail env isn't configured (hydration then no-ops). */
-function attachmentScreeningConfig(): ScreeningConfig | null {
-  if (!ATTACHMENTS_BUCKET || !GUARDRAIL_ID || !GUARDRAIL_VERSION) return null;
-  if (!sharedAttachmentsBedrock) sharedAttachmentsBedrock = new BedrockRuntimeClient({});
-  return { bedrockClient: sharedAttachmentsBedrock, guardrailId: GUARDRAIL_ID, guardrailVersion: GUARDRAIL_VERSION };
-}
-
 /** Terminal task statuses that the reconciler reacts to. */
 const TERMINAL: ReadonlySet<TaskStatusType> = new Set<TaskStatusType>([
   TaskStatus.COMPLETED,
@@ -1417,662 +1367,6 @@ async function resolvePrNumber(taskId?: string): Promise<number | null> {
  * stream retries the batch (idempotent replay is safe). Non-terminal /
  * non-orchestration records are skipped cheaply.
  */
-/**
- * Agent-native planning: a terminal ``coding/decompose-v1`` PLANNING task,
- * extracted from a TaskTable stream record. Distinct from {@link TerminalTaskEvent}
- * — it carries no orchestration_id (it CREATES the graph) and instead carries the
- * decompose mode + caps + parent + the plan-artifact URI (all set by the webhook
- * in channel_metadata / by the agent's deliver_artifact).
- */
-interface DecomposePlanEvent {
-  readonly taskId: string;
-  readonly status: TaskStatusType;
-  readonly parentIssueId: string;
-  readonly workspaceId: string;
-  readonly repo: string;
-  readonly projectId: string;
-  readonly platformUserId: string;
-  readonly mode: 'decompose' | 'auto';
-  readonly maxSubIssues: number;
-  readonly decomposeAllowed: boolean;
-  /**
-   * The label that triggered this decomposition, as resolved from the project
-   * mapping (``label_filter``, default ``bgagent``). Carried through so the
-   * retry hint on the epic panel names the label the operator actually
-   * configured — telling someone to "re-apply `bgagent`" when their project
-   * triggers on something else is a dead end. Absent for a task stamped by an
-   * older webhook, in which case the renderer falls back to the default.
-   */
-  readonly triggerLabel?: string;
-  readonly maxParentBudgetUsd?: number;
-  readonly artifactUri?: string;
-  /**
-   * The planning task's own task_description (the issue title+body — see the
-   * webhook's buildDecompositionTaskDescription). Reused as the single-task
-   * fallback description when the agent declines to decompose (the reconciler
-   * doesn't re-fetch the Linear issue).
-   */
-  readonly taskDescription?: string;
-  /**
-   * Revise loop: 0/absent = original proposal; N≥1 = the Nth re-plan from
-   * reviewer feedback. On a revision the reconciler REPLACES the pending plan
-   * (upsert) instead of create-once, and the proposal renders "round N".
-   */
-  readonly revisionRound?: number;
-  /**
-   * Revise-in-place: the reviewer's FEEDBACK comment id (rides in
-   * channel_metadata so it's task-scoped). The webhook put 👀 on it as the "on it"
-   * ack; when the revised plan lands the reconciler swaps 👀 → ✅ on this SAME
-   * comment so the reviewer can tell it finished (an unsettled 👀 reads as
-   * stuck). Absent on round 0.
-   */
-  readonly revisingFeedbackCommentId?: string;
-}
-
-/**
- * Detect + extract a terminal decompose-planning task from a stream record.
- * Returns null for anything that isn't a coding/decompose-v1 task (the common
- * case — the reconciler is the whole table's consumer). Keyed on the resolved
- * workflow id in the NewImage; the decompose_* channel_metadata (set by the
- * webhook) carries the mode/caps/parent so we act without re-deriving them.
- */
-export function parseDecomposePlanRecord(record: DynamoDBRecord): DecomposePlanEvent | null {
-  const img = record.dynamodb?.NewImage;
-  if (!img) return null;
-  const taskId = img.task_id?.S;
-  const status = img.status?.S as TaskStatusType | undefined;
-  if (!taskId || !status || !TERMINAL.has(status)) return null;
-
-  // Only coding/decompose-v1 tasks. resolved_workflow persists as a MAP.
-  const workflowId = img.resolved_workflow?.M?.id?.S;
-  if (workflowId !== 'coding/decompose-v1') return null;
-
-  const cm = img.channel_metadata?.M;
-  const parentIssueId = cm?.decompose_parent_issue_id?.S ?? cm?.linear_issue_id?.S;
-  const workspaceId = cm?.linear_workspace_id?.S;
-  const projectId = cm?.linear_project_id?.S;
-  const mode = cm?.decompose_mode?.S as ('decompose' | 'auto' | undefined);
-  if (!parentIssueId || !workspaceId || !projectId || (mode !== 'decompose' && mode !== 'auto')) {
-    logger.warn('Decompose plan task terminal but missing routing metadata — skipping', {
-      task_id: taskId, parent_issue_id: parentIssueId, mode,
-    });
-    return null;
-  }
-  const platformUserId = img.user_id?.S ?? '';
-  const repo = img.repo?.S ?? '';
-  // Cap defaults MUST match readProjectCaps / DEFAULT_MAX_SUB_ISSUES (8) so a
-  // task whose caps weren't stamped (older webhook) gates the same as the flow.
-  const maxSubIssues = Number(cm?.decompose_caps_max_sub_issues?.S ?? '8');
-  const decomposeAllowed = (cm?.decompose_caps_allowed?.S ?? 'true') === 'true';
-  const budgetStr = cm?.decompose_caps_max_parent_budget_usd?.S;
-  const artifactUri = img.artifact_uri?.S;
-  const taskDescription = img.task_description?.S;
-  const revisionRoundStr = cm?.decompose_revision_round?.S;
-  const revisionRound = revisionRoundStr !== undefined && Number.isFinite(Number(revisionRoundStr))
-    ? Number(revisionRoundStr) : undefined;
-  const revisingFeedbackCommentId = cm?.decompose_revising_feedback_comment_id?.S;
-  // Blank/whitespace is treated as absent so the renderer's default applies,
-  // rather than rendering an empty label in the retry hint.
-  const triggerLabel = cm?.decompose_trigger_label?.S?.trim() || undefined;
-
-  return {
-    taskId,
-    status,
-    parentIssueId,
-    workspaceId,
-    repo,
-    projectId,
-    platformUserId,
-    mode,
-    maxSubIssues,
-    decomposeAllowed,
-    ...(budgetStr !== undefined && Number.isFinite(Number(budgetStr)) && { maxParentBudgetUsd: Number(budgetStr) }),
-    ...(artifactUri !== undefined && { artifactUri }),
-    ...(taskDescription !== undefined && { taskDescription }),
-    ...(revisionRound !== undefined && { revisionRound }),
-    ...(revisingFeedbackCommentId !== undefined && { revisingFeedbackCommentId }),
-    ...(triggerLabel !== undefined && { triggerLabel }),
-  };
-}
-
-/** Fetch the plan artifact (the agent's plan JSON at artifacts/{task_id}/result.md). */
-async function fetchPlanArtifact(evt: DecomposePlanEvent): Promise<string | null> {
-  if (!ARTIFACTS_BUCKET) {
-    logger.warn('Decompose plan: ARTIFACTS_BUCKET_NAME unset — cannot read plan', { task_id: evt.taskId });
-    return null;
-  }
-  // Prefer the URI the agent recorded; fall back to the conventional key.
-  let bucket = ARTIFACTS_BUCKET;
-  let key = `artifacts/${evt.taskId}/result.md`;
-  if (evt.artifactUri?.startsWith('s3://')) {
-    const rest = evt.artifactUri.slice('s3://'.length);
-    const slash = rest.indexOf('/');
-    if (slash > 0) { bucket = rest.slice(0, slash); key = rest.slice(slash + 1); }
-  }
-  try {
-    const res = await s3().send(new GetObjectCommand({ Bucket: bucket, Key: key }));
-    const body = await res.Body?.transformToString();
-    return body ?? null;
-  } catch (err) {
-    logger.error('Decompose plan: failed to read plan artifact from S3', {
-      task_id: evt.taskId, bucket, key, error: err instanceof Error ? err.message : String(err),
-    });
-    return null; // nosemgrep: ts-silent-success-masking -- best-effort plan read; null → reconcileDecomposePlan posts the planner-error note (honest fallback, not a masked success)
-  }
-}
-
-/**
- * Agent-native planning terminal handler. A coding/decompose-v1 task
- * finished: read its plan artifact, then run the SAME caps → propose/seed tail
- * the inline webhook planner used ({@link applyDecompositionResult}) — either
- * PROPOSE the plan (``:decompose`` → comment + pending plan awaiting
- * ``@bgagent approve``) or SEED immediately (``:auto`` → write back sub-issues +
- * run the graph). The only new thing vs the old planner is WHERE the plan came
- * from (an agent artifact, planned in a real clone with full repo context,
- * instead of a blind 30s Lambda Bedrock call). On a ``seed`` result we release roots +
- * post the panel via the reconciler's own primitives (mirrors the webhook's
- * seedAndReleaseFromGraph). Never throws — the handler loop treats a decompose
- * event as processed regardless.
- */
-async function reconcileDecomposePlan(evt: DecomposePlanEvent): Promise<void> {
-  if (!WORKSPACE_REGISTRY_TABLE) {
-    // No per-workspace token source → can't post the proposal/note or write back.
-    // (Same gate the webhook's decompose path uses; without it there is nothing to do.)
-    logger.warn('Decompose plan terminal but workspace registry table unset — skipping', {
-      task_id: evt.taskId,
-    });
-    return;
-  }
-  const registryTable = WORKSPACE_REGISTRY_TABLE;
-
-  // Idempotency: the TaskTable stream is at-least-once AND the agent writes the
-  // terminal row several times (status, then artifact_uri, then cost/duration —
-  // each a MODIFY that re-delivers this terminal event). Without a claim we'd
-  // re-run the whole handler per delivery and post a fresh proposal comment each
-  // time (observed in practice: 3 duplicate proposals on one issue; the
-  // pending-plan create-once gate kept STATE correct but not the comment). Claim once per
-  // planning task_id — the create-once conditional write means only the first
-  // delivery proceeds; every replay no-ops here. (The webhook's inline path was
-  // shielded by its 60s dedup table; the reconciler has no such upstream gate.)
-  const claimOrchestrationId = deriveOrchestrationId(evt.parentIssueId);
-  const claimed = await claimCommentAck(
-    ddb, ORCHESTRATION_TABLE, claimOrchestrationId, `decompose#${evt.taskId}`,
-    new Date().toISOString(), Math.floor(Date.now() / 1000) + PENDING_PLAN_TTL_SECONDS,
-  );
-  if (!claimed) {
-    logger.info('Decompose plan already reconciled for this task — skipping redelivery', {
-      task_id: evt.taskId, parent_issue_id: evt.parentIssueId,
-    });
-    return;
-  }
-
-  // A decompose-planning task has no seeded orchestration yet, so there's no
-  // stored row to read the surface from — but only the Linear trigger dispatches
-  // planning today, so Linear is the surface by provenance, not by assumption.
-  // (When another surface can dispatch planning, the event should carry its
-  // channel and this becomes a lookup like the loaded-row paths above.)
-  const channel = channelFor('linear');
-  if (!channel) return;
-  // Forward an optional existingCommentId so the flow can EDIT the plan comment
-  // in place on a revision (vs. posting a fresh one).
-  const postComment = async (issueId: string, body: string, existingCommentId?: string): Promise<string | null> => {
-    const ref = await channel.upsertComment(
-      issueRef(issueId, evt.workspaceId), body,
-      existingCommentId ? { commentId: existingCommentId } : undefined,
-    );
-    // Report no id when the surface can't hand one back — a blank id would be
-    // persisted and later addressed as if it were a real comment.
-    return ref?.commentId || null;
-  };
-
-  // Planning RUN did not complete (session failed to start — e.g. a compute
-  // substrate error — cancelled, etc.). Post the honest "couldn't plan, nothing
-  // started" note. We do NOT auto-create a single task here: the webhook
-  // dispatched planning, and nothing has run. renderDecomposeUnavailableNote is
-  // used (NOT renderPlannerErrorNote, which claims "running as a single task" —
-  // false here, nothing ran).
-  if (evt.status !== TaskStatus.COMPLETED) {
-    logger.info('Decompose planning run did not complete — posting decompose-unavailable note', {
-      task_id: evt.taskId, status: evt.status,
-    });
-    await postComment(evt.parentIssueId, renderDecomposeUnavailableNote());
-    return;
-  }
-
-  const planText = await fetchPlanArtifact(evt);
-  if (!planText) {
-    // Run completed but produced no readable plan artifact — again nothing was
-    // started, so the "unavailable, nothing run" note (not "running as single").
-    await postComment(evt.parentIssueId, renderDecomposeUnavailableNote());
-    return;
-  }
-
-  // Parse + validate the agent's plan JSON with the SAME helper the inline
-  // planner's decomposer output flowed through (markdown-fence tolerant, <2-node
-  // collapse to single_task, DAG-validated). Produces a DecompositionResult the
-  // shared tail consumes exactly as it consumes planDecomposition's output.
-  const parsed = parseDecomposerResponse(planText, evt.maxSubIssues, '');
-
-  // Resolve the workspace token once — the :auto write-back needs a GraphQL
-  // transport. A resolution failure only matters for the auto (write-back) path;
-  // the manual proposal path posts via upsertStatusComment (own token) + persists
-  // a pending plan (approve resolves its own token later).
-  const resolved = await resolveLinearOauthToken(evt.workspaceId, registryTable);
-  if (evt.mode === 'auto' && parsed.kind === 'plan' && !resolved) {
-    logger.warn('Decompose :auto: could not resolve OAuth token for write-back', {
-      parent_issue_id: evt.parentIssueId,
-    });
-    // Token unresolved → we can't write back the sub-issues; nothing started.
-    await postComment(evt.parentIssueId, renderDecomposeUnavailableNote());
-    return;
-  }
-
-  const caps: ProjectDecompositionCaps = {
-    decompose_allowed: evt.decomposeAllowed,
-    max_sub_issues: evt.maxSubIssues,
-    ...(evt.maxParentBudgetUsd !== undefined && { max_parent_budget_usd: evt.maxParentBudgetUsd }),
-  };
-
-  // On a REVISION, read the prior plan proposal's comment
-  // id so applyDecompositionResult edits THAT comment in place instead of stacking
-  // a fresh "Updated breakdown". Only on a revision (round 0 has nothing to edit);
-  // best-effort (a missing row just falls back to a fresh post).
-  let priorProposalCommentId: string | undefined;
-  if (evt.revisionRound !== undefined && evt.revisionRound > 0) {
-    const existing = await getPendingPlan(ddb, ORCHESTRATION_TABLE, evt.parentIssueId);
-    priorProposalCommentId = existing?.proposal_comment_id;
-  }
-
-  const result = await applyDecompositionResult({
-    parentIssueId: evt.parentIssueId,
-    planned: parsed,
-    // The agent planned with FULL repo context (the whole point of planning in a
-    // real clone), so a decline is trusted — there is no repo-blindness left to
-    // compensate for, so never route to the "underspecified, ask for detail"
-    // branch here.
-    underspecified: false,
-    caps,
-    autoRun: evt.mode === 'auto',
-    ...(evt.revisionRound !== undefined && { revisionRound: evt.revisionRound }),
-    ...(priorProposalCommentId !== undefined && { priorProposalCommentId }),
-    // Single-task gate: the parent's task_description, so a MANUAL
-    // (``:decompose``) decline can PROPOSE the single task (persist a
-    // pending_kind:'single' plan + wait for approve) instead of auto-running it.
-    ...(evt.taskDescription !== undefined && { singleTaskDescription: evt.taskDescription }),
-    effects: {
-      postComment,
-      putPendingPlan: async ({ nodes, proposalCommentId, revisionRound, repoDigest, repoDigestSha, pendingKind, singleTaskDescription }) => {
-        const row = {
-          ddb,
-          tableName: ORCHESTRATION_TABLE,
-          parentLinearIssueId: evt.parentIssueId,
-          linearWorkspaceId: evt.workspaceId,
-          repo: evt.repo,
-          ...(evt.projectId && { linearProjectId: evt.projectId }),
-          nodes,
-          platformUserId: evt.platformUserId,
-          ...(proposalCommentId !== undefined && { proposalCommentId }),
-          ...(revisionRound !== undefined && { revisionRound }),
-          // Persist the agent's repo digest + its sha so the next revise run
-          // reuses the exploration.
-          ...(repoDigest !== undefined && { repoDigest }),
-          ...(repoDigestSha !== undefined && { repoDigestSha }),
-          // Single-task gate: mark a 'single' pending plan + carry the
-          // task_description approve will run.
-          ...(pendingKind !== undefined && { pendingKind }),
-          ...(singleTaskDescription !== undefined && { singleTaskDescription }),
-          now: new Date().toISOString(),
-          ttlEpochSeconds: Math.floor(Date.now() / 1000) + PENDING_PLAN_TTL_SECONDS,
-        };
-        // Revise loop: a revision (round ≥ 1) must OVERWRITE the prior
-        // pending plan — create-once would silently keep the stale one and
-        // approve would seed it. Round 0 stays create-once (redelivery-safe).
-        return (revisionRound !== undefined && revisionRound > 0)
-          ? replacePendingPlan(row)
-          : putPendingPlan(row);
-      },
-      // Only used on the :auto path; a null token there was already handled above.
-      graphql: linearGraphqlFn(resolved?.accessToken ?? ''),
-    },
-  });
-
-  // The revised plan has now matured (edited in place, or a collapse/over-cap
-  // note posted), so settle the reviewer's FEEDBACK comment 👀 → ✅ — that's how
-  // they can tell the re-plan finished (the 👀 put on it at dispatch previously
-  // never settled → read as stuck). The ONE plan comment updated in place; their
-  // feedback comment now shows ✅. Best-effort — never throws out of the
-  // reconcile. Only on a revision that carried the id.
-  if (evt.revisingFeedbackCommentId) {
-    try {
-      await channel.replaceCommentReaction?.(
-        { commentId: evt.revisingFeedbackCommentId },
-        issueRef(evt.parentIssueId, evt.workspaceId),
-        'succeeded',
-      );
-    } catch (err) {
-      logger.warn('Failed to settle plan-feedback comment 👀→✅ (non-fatal)', {
-        parent_issue_id: evt.parentIssueId, error: err instanceof Error ? err.message : String(err),
-      });
-    }
-  }
-
-  if (result.kind === 'seed') {
-    // :auto wrote back real Linear sub-issues → seed the executor + release roots.
-    // Pass the agreed plan nodes + the proposal comment id so
-    // the seed path can FREEZE the proposal into an "Approved plan" reference +
-    // sweep the started ack, converging on the same shape as the approve path.
-    await seedDecomposedGraph(
-      evt, result.children, resolved!.oauthSecretArn, resolved!.workspaceSlug,
-      {
-        planNodes: parsed.kind === 'plan' ? parsed.plan.nodes : [],
-        ...(result.proposalCommentId !== undefined && { proposalCommentId: result.proposalCommentId }),
-      },
-      resolved!.accessToken,
-    );
-    return;
-  }
-  if (result.kind === 'single_task') {
-    // A REVISION that collapses to one unit must NOT auto-dispatch a coding task:
-    // evt.taskDescription here is the revision META-PROMPT ("You previously
-    // proposed a decomposition… REVISE…"), so createTaskCore would spawn a
-    // nonsensical run (observed in practice). The reviewer explicitly opted
-    // into planning and their feedback merged it into one unit — post an honest
-    // note and let them decide (approve → run as single, or give more feedback);
-    // don't silently burn a run from a meta-prompt.
-    if (evt.revisionRound !== undefined) {
-      logger.info('Decompose revision collapsed to single unit — awaiting user decision (no auto-run)', {
-        parent_issue_id: evt.parentIssueId, revision_round: evt.revisionRound,
-      });
-      await postComment(evt.parentIssueId, renderRevisionToSingleNote());
-      return;
-    }
-    // Round-0 decline: evt.taskDescription is the issue's own title+body, so
-    // running it as a single task is correct — applyDecompositionResult already
-    // posted the decline note; create the task so the work still happens.
-    logger.info('Decompose planner declined — creating single task', {
-      parent_issue_id: evt.parentIssueId, reason: result.reason,
-    });
-    // Silent :auto window: thread the FULL Linear OAuth metadata — without it,
-    // testing found a ~9.5-min "zero output" run on the :auto single-task path.
-    // Without linear_oauth_secret_arn / linear_workspace_slug
-    // the agent can't authenticate to Linear, so it never posts "🤖 Starting",
-    // never transitions state, and never reacts — the run is a total black box
-    // until the PR lands. The secret arn rotates, so we pass the
-    // freshly-resolved one, not a stored id.
-    //
-    // When the planner DECLINES to decompose, this collapsed coding
-    // task must still see the parent's attachments — otherwise a paperclip spec
-    // on a :auto/:decompose issue is lost exactly as it was on the child path.
-    // Fail-closed: an unscreenable attachment rejects the task with a
-    // comment rather than running it blind. (No attachments → returns [].)
-    let singleTaskAttachments: import('./shared/types').PassedAttachmentRecord[] = [];
-    if (resolved?.accessToken) {
-      try {
-        singleTaskAttachments = await hydrateParentAttachmentsForSeed(evt, resolved.accessToken);
-      } catch (err) {
-        if (err instanceof LinearAttachmentError) {
-          logger.warn('Decompose single-task: parent attachment could not be screened — not creating task (fail-closed)', {
-            parent_issue_id: evt.parentIssueId, error: err.message,
-          });
-          await postComment(
-            evt.parentIssueId,
-            `❌ ABCA couldn't safely process an attachment on this issue: ${err.message} ` +
-            'Fix or remove the attachment and re-apply the trigger label.',
-          );
-          return;
-        }
-        throw err;
-      }
-    }
-    await createTaskCore(
-      {
-        repo: evt.repo,
-        task_description: evt.taskDescription ?? `Implement ${evt.parentIssueId}`,
-      },
-      {
-        userId: evt.platformUserId,
-        channelSource: 'linear',
-        channelMetadata: {
-          linear_issue_id: evt.parentIssueId,
-          linear_workspace_id: evt.workspaceId,
-          linear_project_id: evt.projectId,
-          ...(resolved?.oauthSecretArn && { linear_oauth_secret_arn: resolved.oauthSecretArn }),
-          ...(resolved?.workspaceSlug && { linear_workspace_slug: resolved.workspaceSlug }),
-        },
-        ...(singleTaskAttachments.length > 0 && { preScreenedAttachments: singleTaskAttachments }),
-      },
-      `decompose-single-${evt.taskId}`.slice(0, MAX_IDEMPOTENCY_KEY_LENGTH),
-    );
-    return;
-  }
-  // 'handled' (awaiting approval / over-cap / write-back error) or 'noop'
-  // (redelivery) → nothing more; a comment was already posted.
-  logger.info('Decompose plan reconciled', { parent_issue_id: evt.parentIssueId, result: result.kind, reason: result.reason });
-
-  // The plan is now POSTED and AWAITING the reviewer's approve — nothing is
-  // running. The issue was moved to a running state so the board showed the
-  // ~1-2 min planning was happening (round 0: the webhook's dispatch flip; a
-  // revise that ESCALATES to this reconciler: the webhook's escalation flip);
-  // now that it's just a pending plan, a running state would mislead ("looks
-  // like work started while it's awaiting your go"). Revert it to a not-started
-  // state whenever the plan is HANDLED (awaiting approval / single-task proposal
-  // / over-cap) — NOT 'seed' (auto-apply → real work is starting, stay running).
-  //
-  // This fires for BOTH round-0 AND escalated-revise rounds: only the ESCALATED
-  // revise reaches this reconciler (the deterministic interpret→apply revise settles
-  // inline in the webhook with no agent task, so it never flips state and never
-  // arrives here), and that path DID flip to running, so it needs the symmetric
-  // revert. No board flicker — one flip up at dispatch, one flip down here.
-  // approve → the seed path moves it back to running. The revert is guarded +
-  // best-effort per-surface (only demotes an issue still in the state we set).
-  if (result.kind === 'handled') {
-    try {
-      await channel.revertState?.(issueRef(evt.parentIssueId, evt.workspaceId));
-    } catch (err) {
-      logger.warn('Failed to revert issue to not-started after a pending plan (non-fatal)', {
-        parent_issue_id: evt.parentIssueId, error: err instanceof Error ? err.message : String(err),
-      });
-    }
-  }
-}
-
-/**
- * Fetch + screen + store the parent issue's attachments at seed time,
- * so every child of the decomposed epic inherits them. Covers native paperclip
- * attachments (from the context probe) and description-embedded uploads.linear.app
- * links (from the planning task's description = the issue title+body).
- *
- * FAIL-CLOSED: this THROWS `LinearAttachmentError` — the caller
- * rejects the epic rather than seeding children blind — in ANY of three cases:
- *   1. the context probe FAILED (ok:false: 500/timeout) — an empty paperclip list
- *      then means "unknown", not "none", so a paperclip-only spec could be
- *      present-but-unread (this path used to ignore probe.ok and seed anyway);
- *   2. the issue HAS attachments but screening isn't configured;
- *   3. a selected attachment can't be fetched/screened.
- * Returns `[]` ONLY when the probe succeeded AND there's genuinely nothing to
- * hydrate — not a swallowed failure. Returns `passed` records referencing S3.
- */
-async function hydrateParentAttachmentsForSeed(
-  evt: DecomposePlanEvent,
-  accessToken: string,
-): Promise<import('./shared/types').PassedAttachmentRecord[]> {
-  const probe = await probeLinearIssueContext(accessToken, evt.parentIssueId);
-  if (probe.ok === false) {
-    // Case 1: probe failure → fail closed. We can't see native paperclips, so a
-    // paperclip-only spec would vanish; don't seed the epic's children blind.
-    throw new LinearAttachmentError(
-      "ABCA couldn't read this issue's attachments from Linear (the API errored or timed out). "
-      + 'Re-apply the trigger label to retry rather than seed sub-issues on a possibly-missing spec.',
-    );
-  }
-  const paperclips = (probe.attachments ?? []).filter((a) => isLinearUploadsUrl(a.url));
-  const descHasUploads = Boolean(evt.taskDescription?.includes('uploads.linear.app'));
-  if (paperclips.length === 0 && !descHasUploads) return []; // nothing to hydrate
-
-  const screeningConfig = attachmentScreeningConfig();
-  if (!screeningConfig || !ATTACHMENTS_BUCKET) {
-    // Case 2: the issue HAS attachments but we can't screen them — fail closed
-    // (mirrors the webhook single-task path's "screening not configured" reject).
-    throw new LinearAttachmentError(
-      'This issue has attachments, but ABCA attachment screening is not configured. Contact your ABCA admin.',
-    );
-  }
-  // A screening/fetch failure here PROPAGATES (LinearAttachmentError) so the
-  // caller fails the epic closed — we do NOT swallow it to [].
-  return downloadScreenAndStoreLinearAttachments(
-    evt.taskDescription,
-    10, // full per-task budget — the epic parent has no competing public-URL images here
-    {
-      s3Client: s3(),
-      bucketName: ATTACHMENTS_BUCKET,
-      screeningConfig,
-      userId: evt.platformUserId,
-      // Key parent attachments under a stable per-epic id so children share them.
-      taskId: `epic-${evt.parentIssueId}`,
-      accessToken,
-      linearWorkspaceId: evt.workspaceId,
-    },
-    paperclips,
-  );
-}
-
-/**
- * Seed the orchestration from a decompose plan's written-back children +
- * release roots. Mirrors the webhook's seedAndReleaseFromGraph, using the
- * reconciler's own primitives (discoverOrchestration over a declarativeGraphSource
- * → releaseReadyChildren → panel).
- */
-async function seedDecomposedGraph(
-  evt: DecomposePlanEvent,
-  children: readonly SubIssueNode[],
-  oauthSecretArn: string,
-  workspaceSlug: string,
-  cleanup?: { planNodes: readonly PlannedSubIssue[]; proposalCommentId?: string },
-  accessToken?: string,
-): Promise<void> {
-  // Hydrate the parent issue's attachments ONCE here so every child
-  // inherits them (screened + stored to S3, referenced read-only by each child).
-  // FAIL-CLOSED: if the issue has an attachment that can't be safely
-  // screened, do NOT seed children blind to a spec they may need — reject the epic
-  // with a comment, mirroring the single-task path. (No attachments → returns [].)
-  //
-  // REPLAY-SAFE: only hydrate when the orchestration ISN'T already
-  // seeded. The upload key is stable (epic-<parentIssueId>), so a re-hydrate on
-  // redelivery would write a NEW S3 version while the already-persisted meta row
-  // still pins the OLD version_id — the old version then expires under the bucket's
-  // noncurrent-version lifecycle and children fail to download it. The first seed's
-  // records are authoritative; a replay skips straight to release with them.
-  const existingSnapshot = await loadOrchestration(ddb, ORCHESTRATION_TABLE, deriveOrchestrationId(evt.parentIssueId));
-  let preScreenedAttachments: import('./shared/types').PassedAttachmentRecord[] = [];
-  if (accessToken && !existingSnapshot) {
-    try {
-      preScreenedAttachments = await hydrateParentAttachmentsForSeed(evt, accessToken);
-    } catch (err) {
-      if (err instanceof LinearAttachmentError) {
-        logger.warn('Decompose seed: parent attachment could not be screened — NOT seeding children (fail-closed)', {
-          parent_issue_id: evt.parentIssueId, error: err.message,
-        });
-        await channelFor('linear')?.upsertComment(
-          issueRef(evt.parentIssueId, evt.workspaceId),
-          `❌ ABCA couldn't safely process an attachment on this issue: ${err.message} ` +
-          'Fix or remove the attachment and re-apply the trigger label to decompose it.',
-        );
-        return;
-      }
-      throw err; // unexpected — let the handler's error path deal with it
-    }
-  }
-
-  const releaseContext: OrchestrationReleaseContext = {
-    platform_user_id: evt.platformUserId,
-    channel_source: 'linear',
-    linear_oauth_secret_arn: oauthSecretArn,
-    linear_workspace_slug: workspaceSlug,
-    linear_project_id: evt.projectId,
-    // Persisted on the meta row so the panel's retry hint can name the project's
-    // own trigger label. Without this the store/hydrate/render path exists but
-    // every epic renders the default, regardless of configuration.
-    ...(evt.triggerLabel !== undefined && { trigger_label: evt.triggerLabel }),
-    ...(preScreenedAttachments.length > 0 && { pre_screened_attachments: preScreenedAttachments }),
-  };
-  const discovery = await discoverOrchestration({
-    ddb,
-    tableName: ORCHESTRATION_TABLE,
-    parentIssueRef: evt.parentIssueId,
-    credentialsRef: evt.workspaceId,
-    repo: evt.repo,
-    now: new Date().toISOString(),
-    releaseContext,
-    graphSource: declarativeGraphSource(children),
-  });
-  if (discovery.kind !== 'seeded') {
-    logger.info('Decompose :auto seed: discovery non-seeded', { parent_issue_id: evt.parentIssueId, kind: discovery.kind });
-    return;
-  }
-  const snapshot = await loadOrchestration(ddb, ORCHESTRATION_TABLE, discovery.orchestrationId);
-  if (snapshot) {
-    const budget = USER_CONCURRENCY_TABLE
-      ? await readConcurrencyBudget(ddb, USER_CONCURRENCY_TABLE, snapshot.meta.release_context.platform_user_id, MAX_CONCURRENT)
-      : undefined;
-    await releaseReadyChildren(
-      ddb, ORCHESTRATION_TABLE, snapshot.children, snapshot.meta.release_context,
-      createTaskCore, new Date().toISOString(), snapshot.children, 'main', budget,
-    );
-  }
-  // The rows were just seeded with channel_source 'linear' above, so that's the
-  // surface to post the panel on.
-  const channel = channelFor('linear');
-  if (channel) {
-    const parent = issueRef(evt.parentIssueId, evt.workspaceId);
-    try {
-      const fresh = await loadOrchestration(ddb, ORCHESTRATION_TABLE, discovery.orchestrationId);
-      if (fresh) {
-        const commentId = await upsertEpicPanel({
-          channel,
-          parent,
-          children: fresh.children,
-          inProgress: true,
-          mirrorParentState: true,
-        });
-        if (commentId) await setStatusCommentId(ddb, ORCHESTRATION_TABLE, discovery.orchestrationId, commentId);
-      }
-    } catch (err) {
-      logger.warn('Decompose :auto seed: panel post failed (non-fatal)', {
-        parent_issue_id: evt.parentIssueId, error: err instanceof Error ? err.message : String(err),
-      });
-    }
-    // Freeze the auto-applied proposal comment into a static "Approved plan"
-    // reference + sweep the started ack, so the thread converges on the SAME
-    // shape as the approve path (frozen reference + live panel). Best-effort;
-    // cleanup failure is cosmetic. Only when we have the proposal id to freeze.
-    if (cleanup) {
-      try {
-        if (cleanup.proposalCommentId) {
-          await channel.upsertComment(
-            parent,
-            renderApprovedPlanReference(
-              { shouldDecompose: true, reasoning: '', nodes: cleanup.planNodes },
-              evt.revisionRound !== undefined ? { revisionRound: evt.revisionRound } : {},
-            ),
-            { commentId: cleanup.proposalCommentId },
-          );
-        }
-        await channel.sweepNotes?.(
-          parent,
-          cleanup.proposalCommentId ? { commentId: cleanup.proposalCommentId } : undefined,
-        );
-      } catch (err) {
-        logger.warn('Decompose :auto seed: plan-thread cleanup failed (non-fatal)', {
-          parent_issue_id: evt.parentIssueId, error: err instanceof Error ? err.message : String(err),
-        });
-      }
-    }
-  }
-  logger.info('Decompose :auto: orchestration seeded from agent plan', {
-    parent_issue_id: evt.parentIssueId, orchestration_id: discovery.orchestrationId, child_count: discovery.childCount,
-  });
-}
-
 export async function handler(event: DynamoDBStreamEvent): Promise<DynamoDBBatchResponse> {
   let processed = 0;
   // Per-record isolation. A thrown record is reported as a
@@ -2082,16 +1376,6 @@ export async function handler(event: DynamoDBStreamEvent): Promise<DynamoDBBatch
   for (const record of event.Records) {
     const seq = record.dynamodb?.SequenceNumber;
     try {
-      // Agent-native planning: a terminal coding/decompose-v1 PLANNING task
-      // isn't an orchestration child (it has no orchestration_id — it CREATES the
-      // graph). Detect + handle it BEFORE parseTerminalTaskRecord, which would
-      // return null on it (no orchestration_id) and drop it silently.
-      const decomposeEvt = parseDecomposePlanRecord(record);
-      if (decomposeEvt) {
-        await reconcileDecomposePlan(decomposeEvt);
-        processed += 1;
-        continue;
-      }
       const evt = parseTerminalTaskRecord(record);
       if (!evt) continue;
       // Restack cascade: an iteration/restack task on a node X (NOT a child-row task)
